Show user name at the system message for user joined
