+++ packages/app/src/components/user-mention/index.tsx
@@ -8,6 +8,7 @@ const Root = styled.span`
   color: ${(props) => props.theme.colors.invertedText};
   border-radius: ${(props) => props.theme.radii.small}px;
   padding: 1px 4px;
+  margin-right: ${(props) => props.theme.space[1]}px;
 `;
 
 export const UserMention: React.FC<{ name: string; userId?: string }> =
